+++ src/api/errands.ts
@@ -3,8 +3,7 @@ import { DELETE, GET, PATCH, POST } from "@utils/axios";
 import {
   ErrandPreviewResponseBody,
   ErrandDetailResponseBody,
-  Resume,
-  Helper,
+  ResumePreview,
 } from "@type/response";
 import { TabType } from "@type/client";
 import { ERREND_REQUEST_SIZE } from "@constant/request";
@@ -77,20 +76,12 @@ export const useErrandDetail = (errandId: string) => {
   return useQuery(["errandDetail"], () => getErrandDetail(errandId));
 };
 
-const getHelperList = async (errandId: string): Promise<Helper[]> => {
+const getHelperList = async (errandId: string): Promise<ResumePreview[]> => {
   const { data } = await GET(`/errand/${errandId}/helpers`);
   return data;
 };
 export const useHelperList = (errandId: string) => {
   return useQuery(["helperList"], () => getHelperList(errandId));
-};
-
-const getHelperDetail = async (helpId: string): Promise<Resume> => {
-  const { data } = await GET(`/help/${helpId}`);
-  return data;
-};
-export const useHelperDetail = (helpId: string) => {
-  return useQuery(["helperDetail", helpId], () => getHelperDetail(helpId));
 };
 
 export const selectHelper = async (errandId: string, applierId: string) => {
