+++ src/api/help.ts
@@ -1,5 +1,15 @@
 import { ApplyHelperRequestBody } from "@type/request";
-import { DELETE, POST } from "@utils/axios";
+import { Resume } from "@type/response";
+import { DELETE, GET, POST } from "@utils/axios";
+import { useQuery } from "react-query";
+
+const getHelperDetail = async (helpId: string): Promise<Resume> => {
+  const { data } = await GET(`/help/${helpId}`);
+  return data;
+};
+export const useHelperDetail = (helpId: string) => {
+  return useQuery(["helperDetail", helpId], () => getHelperDetail(helpId));
+};
 
 export const applyErrand = async (requestBody: ApplyHelperRequestBody) => {
   const { data } = await POST(`/help`, requestBody);
